feat: reset alembic history
