--- backend/alembic/versions/16593effbe7b_add_vector_column_to_professor_.py
@@ -1,46 +1,0 @@
-"""Add vector column to 'professor_educations'
-
-Revision ID: 16593effbe7b
-Revises: 
-Create Date: 2025-01-13 22:09:18.298848
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-import pgvector.sqlalchemy
-
-# revision identifiers, used by Alembic.
-revision: str = '16593effbe7b'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column(
-        'professor_educations',
-        sa.Column(
-            'dense_vector',
-            pgvector.sqlalchemy.vector.VECTOR(dim=1024),
-            nullable=False
-        )
-    )
-    op.add_column(
-        'professor_educations',
-        sa.Column(
-            'sparse_vector',
-            pgvector.sqlalchemy.sparsevec.SPARSEVEC(dim=250002),
-            nullable=False
-        )
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('professor_educations', 'sparse_vector')
-    op.drop_column('professor_educations', 'dense_vector')
-    # ### end Alembic commands ###
