+++ backend/alembic/versions/9ded957204fe_add_cascade_notices.py
@@ -1,0 +1,36 @@
+"""add cascade notices
+
+Revision ID: 9ded957204fe
+Revises: 
+Create Date: 2025-01-18 16:05:59.708780
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+
+
+# revision identifiers, used by Alembic.
+revision: str = '9ded957204fe'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_constraint('notice_attachments_notice_id_fkey', 'notice_attachments', type_='foreignkey')
+    op.create_foreign_key(None, 'notice_attachments', 'notices', ['notice_id'], ['id'], ondelete='CASCADE')
+    op.drop_constraint('notice_content_chunks_notice_id_fkey', 'notice_content_chunks', type_='foreignkey')
+    op.create_foreign_key(None, 'notice_content_chunks', 'notices', ['notice_id'], ['id'], ondelete='CASCADE')
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_constraint(None, 'notice_content_chunks', type_='foreignkey')
+    op.create_foreign_key('notice_content_chunks_notice_id_fkey', 'notice_content_chunks', 'notices', ['notice_id'], ['id'])
+    op.drop_constraint(None, 'notice_attachments', type_='foreignkey')
+    op.create_foreign_key('notice_attachments_notice_id_fkey', 'notice_attachments', 'notices', ['notice_id'], ['id'])
+    # ### end Alembic commands ###
